+++ src/content-scripts/transfer-list/transfer-list-update.ts
@@ -1,5 +1,5 @@
-import {Preset} from "../preset";
-import {Update} from "../update";
+import {Preset} from "../../preset";
+import {Update} from "../../update";
 
 export interface TransferListUpdate {
   selectedPreset: Update<Preset>,
