+++ src/content-scripts/transfer-list/transfer-list.tsx
@@ -1,7 +1,7 @@
-import {el} from "../dom/jsx-runtime";
-import {Preset} from "../preset";
-import {PresetInputType} from "../preset-input-type";
-import {PresetInput} from "../preset-input";
+import {el} from "../../dom/jsx-runtime";
+import {Preset} from "../../preset";
+import {PresetInputType} from "../../preset-input-type";
+import {PresetInput} from "../../preset-input";
 import {TransferListViewModel} from "./transfer-list-view-model";
 import {TransferListUpdate} from "./transfer-list-update";
 import {browser} from "webextension-polyfill-ts";
